requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
discord-protos==0.0.2
discord.py-self==2.0.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
protobuf==5.28.3
python-dotenv==1.0.1
yarl==1.17.1
